% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMethod.R
\name{clusterBoundaryTest}
\alias{clusterBoundaryTest}
\title{Boundary Test for the GapBin Method}
\usage{
clusterBoundaryTest(point, bin, inputs)
}
\arguments{
\item{point}{A point to be tested to see if it's inside the provided bin object}

\item{bin}{A bin object}

\item{inputs}{A list containing additional input parameters required by user supplied functions}
}
\value{
Returns a value of TRUE if the point belongs in the bin, FALSE otherwise
}
\description{
Tests whether a given point is contained in the bin boundary produced by the GapBin method
}
