% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMethod.R
\name{clusterMeasure}
\alias{clusterMeasure}
\title{Compute the Gap Measure of Rahman & Oldford}
\usage{
clusterMeasure(bin, inputs)
}
\arguments{
\item{bin}{An object of class bin}

\item{inputs}{A list containing additional input parameters required by user supplied functions}
}
\value{
\item{measure}{Returns the measure to be associated with the supplied bin object}
}
\description{
\code{gapMeasure} computes the Gap measure of Rahman & Oldford (2018) as a measure of dissimilarity within a bin
}
\references{
Radchenko et. al. Convex Clustering via L1 fusion penalization. JRSS-B (2017).
}
