% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMethod.R
\name{clusterSelect}
\alias{clusterSelect}
\title{Select the Bin to be Split}
\usage{
clusterSelect(bins, inputs)
}
\arguments{
\item{bins}{A list containing bin objects}

\item{inputs}{A list containing additional input parameters required by user supplied functions}
}
\value{
\item{Index}{The index number of the chosen bin to split}
}
\description{
\code{clusterSelect} uses the measures computed using \code{clusterMeasure} to determine which of 
the bins should be chosen for splitting.
}
