% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMethod.R
\name{clusterStop}
\alias{clusterStop}
\title{Stopping Criteria for the Clustered Binning Algorithm}
\usage{
clusterStop(bins, inputs)
}
\arguments{
\item{bins}{A list of objects of class bin}

\item{inputs}{A list containing additional input parameters required by user supplied functions}
}
\value{
Returns TRUE if the algorithm should be stopped, FALSE otherwise
}
\description{
\code{clusterStop} Computes the stopping critera for the Clustered Binning Algorithm.
}
