% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doNextSplit.R
\name{doNextSplit}
\alias{doNextSplit}
\title{Perform one Additional Iteration of treebin}
\usage{
doNextSplit(treebinr_obj, selectBin, splitBin, binMeasure, makePoint, inputs)
}
\arguments{
\item{treebinr_obj}{An object of class treebinr}

\item{selectBin}{A user supplied function for choosing between bins to be split}

\item{splitBin}{A user supplied function for splitting a bin}

\item{binMeasure}{A user supplied function to compute the measure associated with each bin}

\item{makePoint}{A user supplied function to turn the contents of a bin into a single point}

\item{inputs}{A list containing additional input parameters required by user supplied functions}
}
\value{
The return value is an object of class treebinr, which contains the following 
\item{points}{A matrix containing the reduced point configuration}
\item{counts}{A vector containing the number of points in each bin}
\item{bins}{A list containing bin objects, which detail the contents of each bin}
\item{tree}{An undirected graph object for the binning tree}
}
\description{
Given a treebinr object, \code{nextSplit} performs one additional iteration of treebin based on user supplied input function
}
\examples{

}
