% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewMeasure.R
\name{ewMeasure}
\alias{ewMeasure}
\title{Compute the Equal Width Measure}
\usage{
ewMeasure(bin, inputs)
}
\arguments{
\item{bin}{An object of class bin}

\item{inputs}{A list containing additional input parameters required by user supplied functions}
}
\value{
\item{measure}{Returns the measure to be associated with the supplied bin object}
}
\description{
\code{ewMeasure} computes the Equal Width measure as a measure of dissimilarity within a bin
}
