% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewSelect.R
\name{ewSelect}
\alias{ewSelect}
\title{Select the Bin to be Split}
\usage{
ewSelect(bins, inputs)
}
\arguments{
\item{bins}{A list containing bin objects}

\item{inputs}{A list containing additional input parameters required by user supplied functions}
}
\value{
\item{Index}{The index number of the chosen bin to split}
}
\description{
\code{ewSelect} uses the measures computed using \code{ewMeasure} to determine which of 
the bins should be chosen for splitting.
}
