% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewStop.R
\name{ewStop}
\alias{ewStop}
\title{Stopping Criteria for the Equal Width Binning Algorithm}
\usage{
ewStop(bins, inputs)
}
\arguments{
\item{bins}{A list of objects of class bin}

\item{inputs}{A list containing additional input parameters required by user supplied functions}
}
\value{
Returns TRUE if the algorithm should be stopped, FALSE otherwise
}
\description{
\code{ewStop} Computes the stopping critera for the Equal Width Binning Algorithm.
}
