% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapMeasure.R
\name{gapMeasure}
\alias{gapMeasure}
\title{Compute the Gap Measure of Rahman & Oldford}
\usage{
gapMeasure(bin, inputs)
}
\arguments{
\item{bin}{An object of class bin}

\item{inputs}{A list containing additional input parameters required by user supplied functions}
}
\value{
\item{measure}{Returns the measure to be associated with the supplied bin object}
}
\description{
\code{gapMeasure} computes the Gap measure of Rahman & Oldford (2018) as a measure of dissimilarity within a bin
}
\references{
Adam Rahman & Wayne Oldford (2018). Tree-based Binning.
}
