% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapPoints.R
\name{gapPoints}
\alias{gapPoints}
\title{Convert Bins into a Reduced Point Configuration}
\usage{
gapPoints(bin, inputs)
}
\arguments{
\item{bin}{A bin object}

\item{inputs}{A list containing additional input parameters required by user supplied functions}
}
\value{
A single point to be included in the final configuration
}
\description{
Given a bin object, reduce to a single point to make up the final point configuration
}
