% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapSelect.R
\name{gapSelect}
\alias{gapSelect}
\title{Select the Bin to be Split}
\usage{
gapSelect(bins, inputs)
}
\arguments{
\item{bins}{A list containing bin objects}

\item{inputs}{A list containing additional input parameters required by user supplied functions}
}
\value{
\item{Index}{The index number of the chosen bin to split}
}
\description{
\code{gapSelect} uses the measures computed using \code{gapMeasure} to determine which of 
the bins should be chosen for splitting.
}
