% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapSplit.R
\name{gapSplit}
\alias{gapSplit}
\title{Split the Chosen Bin}
\usage{
gapSplit(bin, binMeasure, inputs)
}
\arguments{
\item{bin}{An object of class bin}

\item{binMeasure}{A user supplied function to compute the measure associated with each bin}

\item{inputs}{A list containing additional input parameters required by user supplied functions}
}
\value{
\item{newBins}{A list containing two bin objects to be added to the existing list of bins}
}
\description{
\code{gapSplit} splits the bin chosen by \code{gapSelect} by placing a boundary at the midpoint of the two points 
with the largest gap measure computed by \code{gapMeasure}
}
