% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapStop.R
\name{gapStop}
\alias{gapStop}
\title{Stopping Criteria for the GapBin Algorithm}
\usage{
gapStop(bins, inputs)
}
\arguments{
\item{bins}{A list of objects of class bin}

\item{inputs}{A list containing additional input parameters required by user supplied functions}
}
\value{
Returns TRUE if the algorithm should be stopped, FALSE otherwise
}
\description{
\code{gapStop} Computes the stopping critera for the GapBin algorithm.
}
\references{
Adam Rahman & Wayne Oldford (2018). Tree-based Binning.
}
