% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapUpdate.R
\name{gapUpdate}
\alias{gapUpdate}
\title{Create a Bin from a Number of Pruned Bins}
\usage{
gapUpdate(bins, binMeasure, inputs)
}
\arguments{
\item{bins}{A list of bin objects to be pruned}

\item{binMeasure}{A user supplied function to compute the measure associated with each bin}

\item{inputs}{A list containing additional input parameters required by user supplied functions}
}
\value{
An updated bin object containing the contents of the pruned bins
}
\description{
Given a number of bins that we would like to prune, combine them into a single bin and update all relevant bin information
}
