% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineClasses.R
\docType{class}
\name{treebinr-class}
\alias{treebinr-class}
\alias{treebinr}
\title{The S4 class treebinr to represent the data reduction output from tree binning.}
\description{
The S4 class treebinr to represent the data reduction output from tree binning.
}
\section{Slots}{

\describe{
\item{\code{points}}{A matrix whose rows represent the points after data reduction.  These
are the locations of the bins in the original space.}

\item{\code{counts}}{These are the counts in the bin, or the number of original values that
each point now represents}

\item{\code{bins}}{A vector of bins, one for each of the reduced points.  Each bin itself
an instance of the treebinr class `bin` and so contains much more information}

\item{\code{tree}}{A large (n-1) by (n-1) matrix containing information necessary to reconstruct
the tree}
}}

