% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{treePrune}
\alias{treePrune}
\title{Prune a Bin}
\usage{
treePrune(bin, treebinr_obj, binMeasure, makePoint, updateBin, inputs)
}
\arguments{
\item{bin}{bin An object of class bin (contained in treebinr_obj) to be pruned}

\item{treebinr_obj}{An object of class treebinr}

\item{binMeasure}{A user supplied function to compute the measure associated with each bin}

\item{makePoint}{A user supplied function to turn the contents of a bin into a single point}

\item{updateBin}{A user supplied function to produce a single bin given the pruned bins}

\item{inputs}{A list containing additional input parameters required by user supplied functions}
}
\value{
Returns an updated treebinr object containing the newly pruned bins
\item{points}{A matrix containing the reduced point configuration}
\item{counts}{A vector containing the number of points in each bin}
\item{bins}{A list containing bin objects, which detail the contents of each bin}
\item{tree}{An undirected graph object for the binning tree}
}
\description{
Given a bin, \code{prune} combines all of the bins sister bins into a single node.
}
