% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{treeSplit}
\alias{treeSplit}
\title{Split A Node}
\usage{
treeSplit(bin, treebinr_obj, binMeasure, splitBin, makePoint, inputs)
}
\arguments{
\item{bin}{An object of class bin (contained in treebinr_obj) to be split}

\item{treebinr_obj}{An object of class treebinr}

\item{binMeasure}{A user supplied function to compute the measure associated with each bin}

\item{splitBin}{A user supplied function for splitting a bin}

\item{makePoint}{A user supplied function to turn the contents of a bin into a single point}

\item{inputs}{A list containing additional input parameters required by user supplied functions}
}
\value{
The return value is an object of class treebinr, which contains the following 
\item{points}{A matrix containing the reduced point configuration}
\item{counts}{A vector containing the number of points in each bin}
\item{bins}{A list containing bin objects, which detail the contents of each bin}
\item{tree}{An undirected graph object for the binning tree}
}
\description{
Given a bin, \code{split} splits it into two or more bins according to functions provided by the user
}
