% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimTo.R
\name{treeTrim}
\alias{treeTrim}
\title{Trim the Binning Tree to a Given Number of Bins}
\usage{
treeTrim(num, treebinr_obj, getMeasure = function(bin, inputs) {    
  bin@measure }, binMeasure, makePoint, updateBin, inputs = NULL)
}
\arguments{
\item{num}{The number of points desired in the trimmed configuration}

\item{treebinr_obj}{An object of class treebinr}

\item{getMeasure}{A user supplied function to compute the measure to be compared between bins. By default, returns whatever value is in the measure slot of the bin object.}

\item{binMeasure}{A user supplied function to compute the measure associated with each bin}

\item{makePoint}{A user supplied function to turn the contents of a bin into a single point}

\item{inputs}{A list containing additional input parameters required by user supplied functions}

\item{splitBin}{A user supplied function for splitting a bin}
}
\value{
The return value is an object of class treebinr, which contains the following 
\item{points}{A matrix containing the reduced point configuration}
\item{counts}{A vector containing the number of points in each bin}
\item{bins}{A list containing bin objects, which detail the contents of each bin}
\item{tree}{An undirected graph object for the binning tree}
}
\description{
Given a treebinr object, \code{trimTo} trims the tree to the desired size by removing the bins with the lowest measure value.
}
